Add files via upload
Overview:
We analysed 10 years of U.S. flight data to identify when flights are least likely to be delayed.
Initial analysis highlighted arrival delays as the main contributing factor. After data wrangling, we visualized yearly patterns to see which specific times of day showed the lowest delays.

To give a broader view across the entire 10-year period, scheduled arrival times were grouped into four parts of the day (Night, Morning, Afternoon, Evening), and the percentage of delayed flights was calculated for each period.

To identify the best days of the week, we summed arrival and departure delays, then aggregated the data to find average delays per day. Days of the week were mapped, and a final visualisation was created for indiviual years. Finally we repeat the steps but across the whole 10 year period using a heatmap comparing with the individual delay variables and the combined delay vairable across the days of the week. 

Detailed methodology and calculations are documented in the code.

Key Insights:

Best Time of Day:
While Night flights showed the lowest delays in several individual years, when aggregated across all 10 years, Morning flights had the lowest overall delay percentage. This suggests that the best time to fly differs in short-term trends or long-term patterns.

Best Day of the Week:
Both individual year plots and the 10-year summary indicate that Saturday consistently has the lowest average flight delays.

How to Run:
All relevant packages are loaded, and the analysis is fully implemented in both Python and R.
Each code chunk includes detailed explanations to guide you through the steps — from data cleaning and wrangling to visualisation and insights.
To replicate the analysis, simply follow the code chunks in order within your R or Python environment, ensuring the dataset files are correctly loaded.